+++ create-vite-app/template-vue/main.js
@@ -1,5 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import './index.css'
-
-createApp(App).mount('#app')