+++ src/node/cli.ts
@@ -10,6 +10,7 @@ if (argv.debug) {
 }
 
 import os from 'os'
+import path from 'path'
 import chalk from 'chalk'
 import { UserConfig, resolveConfig } from './config'
 
@@ -94,7 +95,7 @@ async function resolveOptions() {
   // normalize root
   // assumes all commands are in the form of `vite [command] [root]`
   if (argv._[1] && !argv.root) {
-    argv.root = argv._[1]
+    argv.root = path.isAbsolute(argv._[1]) ? argv._[1] : path.resolve(argv._[1])
   }
 
   const userConfig = await resolveConfig(argv.config || argv.c)
@@ -141,12 +142,12 @@ async function runServe(
             type: detail.address.includes('127.0.0.1')
               ? 'Local:   '
               : 'Network: ',
-            ip: detail.address.replace('127.0.0.1', 'localhost')
+            host: detail.address.replace('127.0.0.1', 'localhost')
           }
         })
-        .forEach((address: { type?: String; ip?: String }) => {
-          const url = `http://${address.ip}:${chalk.bold(port)}/`
-          console.log(`  > ${address.type} ${chalk.cyan(url)}`)
+        .forEach(({ type, host }) => {
+          const url = `http://${host}:${chalk.bold(port)}/`
+          console.log(`  > ${type} ${chalk.cyan(url)}`)
         })
     })
     console.log()