+++ src/node/server/serverPluginModuleRewrite.ts
@@ -58,33 +58,32 @@ export const moduleRewritePlugin: ServerPlugin = ({
   const scriptRE = /(<script\b[^>]*>)([\s\S]*?)<\/script>/gm
   const srcRE = /\bsrc=(?:"([^"]+)"|'([^']+)'|([^'"\s]+)\b)/
 
-  async function rewriteIndex(html: string) {
+  async function rewriteIndex(importer: string, html: string) {
     await initLexer
-    let hasInjectedDevFlag = false
-    const importer = '/index.html'
-    return html!.replace(scriptRE, (matched, openTag, script) => {
-      const devFlag = hasInjectedDevFlag ? `` : devInjectionCode
-      hasInjectedDevFlag = true
-      if (script) {
-        return `${devFlag}${openTag}${rewriteImports(
-          root,
-          script,
-          importer,
-          resolver
-        )}</script>`
-      } else {
-        const srcAttr = openTag.match(srcRE)
-        if (srcAttr) {
-          // register script as a import dep for hmr
-          const importee = cleanUrl(
-            slash(path.resolve('/', srcAttr[1] || srcAttr[2]))
-          )
-          debugHmr(`        ${importer} imports ${importee}`)
-          ensureMapEntry(importerMap, importee).add(importer)
+    return (
+      devInjectionCode +
+      html!.replace(scriptRE, (matched, openTag, script) => {
+        if (script) {
+          return `${openTag}${rewriteImports(
+            root,
+            script,
+            importer,
+            resolver
+          )}</script>`
+        } else {
+          const srcAttr = openTag.match(srcRE)
+          if (srcAttr) {
+            // register script as a import dep for hmr
+            const importee = cleanUrl(
+              slash(path.resolve('/', srcAttr[1] || srcAttr[2]))
+            )
+            debugHmr(`        ${importer} imports ${importee}`)
+            ensureMapEntry(importerMap, importee).add(importer)
+          }
+          return matched
         }
-        return `${devFlag}${matched}`
-      }
-    })
+      })
+    )
   }
 
   app.use(async (ctx, next) => {
@@ -94,13 +93,13 @@ export const moduleRewritePlugin: ServerPlugin = ({
       return
     }
 
-    if (ctx.path === '/index.html') {
+    if (ctx.path.endsWith('.html')) {
       let html = await readBody(ctx.body)
       if (html && rewriteCache.has(html)) {
-        debug('/index.html: serving from cache')
+        debug(`${ctx.path}: serving from cache`)
         ctx.body = rewriteCache.get(html)
       } else if (html) {
-        ctx.body = await rewriteIndex(html)
+        ctx.body = await rewriteIndex(ctx.path, html)
         rewriteCache.set(html, ctx.body)
       }
       return
@@ -304,7 +303,8 @@ export const resolveImport = (
     }
 
     // mark non-src imports
-    if (!jsSrcRE.test(pathname)) {
+    const ext = path.extname(pathname)
+    if (ext && !jsSrcRE.test(pathname)) {
       query += `${query ? `&` : `?`}import`
     }